+++ tests/eventbus/test_websocket_manager.py
@@ -16,7 +16,7 @@ from starlette.websockets import WebSocket
 
 from antarest.core.jwt import JWTUser
 from antarest.core.model import PermissionInfo
-from antarest.eventbus.web import ConnectionManager, WebsocketMessage, WebsocketMessageAction
+from antarest.eventbus.connections import ConnectionManager, WebsocketMessage, WebsocketMessageAction
 
 
 class ConnectionManagerTest(IsolatedAsyncioTestCase):
